added places and database reading

+++ app/src/main/java/com/music/cornell/music/Location.java
@@ -5,5 +5,5 @@ package com.music.cornell.music;
  */
 
 public interface Location {
-    public boolean isInLocation(float lat, float lng);
+    public boolean isInLocation(double lat, double lng);
 }
